[lib][sm] Fix UBSAN error
SM_ERR_UNDEFINED_SMC is returned in a long tiggering an overflow on 32
bit kernels.

Bug: 129300035

+++ lib/sm/include/lib/sm/sm_err.h
@@ -26,7 +26,7 @@
 /* Errors from the secure monitor */
 
 /* Unknown SMC (defined by ARM DEN 0028A(0.9.0) */
-#define SM_ERR_UNDEFINED_SMC 0xFFFFFFFF
+#define SM_ERR_UNDEFINED_SMC ((long)0xFFFFFFFF)
 #define SM_ERR_INVALID_PARAMETERS -2
 /* Got interrupted. Call back with restart SMC */
 #define SM_ERR_INTERRUPTED -3
